requirements.txt
kafka-python==2.0.2
pandas==2.0.3
psycopg2-binary==2.9.7
pyspark==3.4.1
requests==2.31.0
python-dotenv==1.0.0